+++ generate.py
@@ -19,6 +19,7 @@ from string import Template
 
 CODEPOINT_CHARACTER_SEQUENCE_FILE_NAME = 'codepoint-character-sequence.txt'
 SEQUENCE_CHARACTERS_FILE_NAME = 'sequence-characters.txt'
+CHARACTERS_UNLIKELY_FILE_NAME = 'characters-unlikely.txt'
 CHARACTERS_TRADITIONAL_FILE_NAME = 'characters-traditional.txt'
 CHARACTERS_SIMPLIFIED_FILE_NAME = 'characters-simplified.txt'
 
@@ -50,6 +51,15 @@ SEQUENCE_CHARACTERS_TEMPLATE = Template(
     f'{STROKE_DATA_NOTICE}\n\n'
     f'{CREATIVE_COMMONS_NOTICE}\n\n'
     f'# Contains tab-separated (stroke sequence, characters) pairs.\n\n'
+    f'{AUTOMATIC_GENERATION_NOTICE}\n\n'
+    f'$body\n'
+)
+
+CHARACTERS_UNLIKELY_TEMPLATE = Template(
+    f'# # {CHARACTERS_UNLIKELY_FILE_NAME}\n\n'
+    f'{STROKE_DATA_NOTICE}\n\n'
+    f'{PUBLIC_DOMAIN_NOTICE}\n\n'
+    f'# Contains characters unlikely to have font support on Android 7.0.\n\n'
     f'{AUTOMATIC_GENERATION_NOTICE}\n\n'
     f'$body\n'
 )
@@ -126,6 +136,7 @@ def to_sequence_set(sequence_regex):
 
 
 def main():
+    unlikely_characters = set()
     traditional_characters = set()
     simplified_characters = set()
     dual_characters = set()
@@ -165,6 +176,9 @@ def main():
             )
             sys.exit(1)
 
+        if font_support == '!':
+            unlikely_characters.add(character)
+
         if character_type == '^':
             traditional_characters.add(character)
         elif character_type == '*':
@@ -186,6 +200,12 @@ def main():
     with open(SEQUENCE_CHARACTERS_FILE_NAME, 'w', encoding='utf-8') as sequence_characters_file:
         sequence_characters_file.write(SEQUENCE_CHARACTERS_TEMPLATE.substitute(body=sequence_characters_lines))
     print(f'Finished generating `{SEQUENCE_CHARACTERS_FILE_NAME}` ({sequence_count} stroke sequences).')
+
+    unlikely_character_count = len(unlikely_characters)
+    unlikely_character_lines = '\n'.join(sorted(unlikely_characters))
+    with open(CHARACTERS_UNLIKELY_FILE_NAME, 'w', encoding='utf-8') as characters_unlikely_file:
+        characters_unlikely_file.write(CHARACTERS_UNLIKELY_TEMPLATE.substitute(body=unlikely_character_lines))
+    print(f'Finished generating `{CHARACTERS_UNLIKELY_FILE_NAME}` ({unlikely_character_count} characters).')
 
     traditional_character_count = len(traditional_characters)
     traditional_character_lines = '\n'.join(sorted(traditional_characters))